Refactor state machine
This is a major refactor of the state machine that simplifies things significantly. It requires fewer nested complex structures, and no branching to handle excluded tasks. It's also more resilant to invalid data in the input message, and the input format is more extenisble when it comes to new task types