Encode plus to %2B for S3 copies
When passing an object to the CopySource of copyObject, if the object key includes a plus character, the SDK/API will only properly handle that if it is encoded to %2B.

The S3 API/SDK expects this value to be URL encoded in a way that preserves forward slashes (among other things). Normally that sort of encoding (URL vs component) would not encoded pluses. Since it's necessary here, the encoding must be done manually, *after* the rest of the string has been URL encoded as usual.