URL decode artifact file names
When a job has an HTTP source, the last portion of the S3 object key for the artifact is derived from the last portion of the URL path.

Previously, that key would include encoded characters (e.g., %20 for a space).

This change will decode characters, to try to get the artifact key to match the original HTTP file.

The artifact key is entierly internal to task execution and opaque to any job output, and is never inherited by things like a copy or transcode task. So this doesn't really change anything, but it does mean there's less variation between input and artifact naming, which can be helpful with debugging.